requirements.txt
streamlit>=1.33,<1.39
pandas>=2.0,<3
numpy>=1.23,<3
matplotlib>=3.7,<4
plotly>=5.18,<6
requests>=2.31,<3
wordcloud>=1.9,<2
pillow>=10,<11
nltk>=3.8,<4
scikit-learn>=1.3,<2
vaderSentiment>=3.3.2,<4
textblob>=0.17,<0.18
python-dotenv>=1.0,<2
psycopg2-binary>=2.9,<3
sib-api-v3-sdk==7.6.0
flask>=2.2,<3
gunicorn>=21,<22